4.4.21: Avoid using zero when season/episode information is not present

+++ src/Tvheadend.cpp
@@ -2543,9 +2543,9 @@ void CTvheadend::ParseRecordingAddOrUpdate ( htsmsg_t *msg, bool bAdd )
 
   /* season/episode/part */
   if (!htsmsg_get_u32(msg, "seasonNumber", &season))
-    rec.SetSeason(season);
+    rec.SetSeason(static_cast<int32_t>(season));
   if (!htsmsg_get_u32(msg, "episodeNumber", &episode))
-    rec.SetEpisode(episode);
+    rec.SetEpisode(static_cast<int32_t>(episode));
   if (!htsmsg_get_u32(msg, "partNumber", &part))
     rec.SetPart(part);
 
@@ -2654,9 +2654,9 @@ bool CTvheadend::ParseEvent ( htsmsg_t *msg, bool bAdd, Event &evt )
   if (!htsmsg_get_s64(msg, "firstAired", &s64))
     evt.SetAired((time_t)s64);
   if (!htsmsg_get_u32(msg, "seasonNumber", &u32))
-    evt.SetSeason(u32);
+    evt.SetSeason(static_cast<int32_t>(u32));
   if (!htsmsg_get_u32(msg, "episodeNumber", &u32))
-    evt.SetEpisode(u32);
+    evt.SetEpisode(static_cast<int32_t>(u32));
   if (!htsmsg_get_u32(msg, "partNumber", &u32))
     evt.SetPart(u32);
   if ((str = htsmsg_get_str(msg, "serieslinkUri")) != NULL)
